pyproject.toml
[project]
name = "projeto-ed-satc"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "pyspark==4.1.0",
    "delta-spark==4.1.0",
    "pyiceberg[pyarrow,pandas]>=0.7.0",
]

[dependency-groups]
dev = [
    "isort>=5.13.2",
    "blue>=0.9.1",
    "pytest>=8.2.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.23",
    "pymdown-extensions>=10.8.1",
    "mkdocstrings>=0.25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
